# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Open-source drug information gateway. Provides NDC lookup, therapeutic class search, drug interactions, RxNorm fuzzy search, and structured product label data — all through a clean REST API with Redis caching, per-key rate limiting, and circuit breaker protection.\n\n## Getting Started\n1. **Create an API key:** `POST /admin/keys` with `Authorization: Bearer {ADMIN_SECRET}` and body `{\"app_name\": \"my-app\", \"origins\": [\"*\"], \"rate_limit\": 1000}`\n2. **Search for a drug:** `GET /v1/drugs/autocomplete?q=lipit` with header `X-API-Key: {your_key}` — returns matching drug names\n3. **Look up details:** `GET /v1/drugs/class?name=atorvastatin` — returns therapeutic classes and brand names\n4. **Get interactions:** `GET /v1/drugs/info?name=warfarin` — returns FDA label interaction warnings\n5. **Check multi-drug interactions:** `POST /v1/drugs/interactions` with `{\"drugs\": [{\"name\": \"warfarin\"}, {\"name\": \"aspirin\"}]}`",
        "title": "drug-gate API",
        "contact": {
            "name": "drug-gate on GitHub",
            "url": "https://github.com/finish06/drug-gate"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/finish06/drug-gate/blob/main/LICENSE"
        },
        "version": "0.9.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/keys": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Returns all provisioned API keys with their metadata including app name, allowed origins, rate limit, active status, and creation timestamp. Use this to audit which keys are active and their configurations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_finish06_drug-gate_internal_apikey.APIKey"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Provisions a new publishable API key (pk_*) with the specified app name, allowed CORS origins, and per-minute rate limit. The full key value is only returned in this response and cannot be retrieved later. Store it securely.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "Key creation parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.createKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_apikey.APIKey"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/keys/{key}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Returns metadata for a specific API key including app name, allowed origins, rate limit, active status, and expiration. Use this to inspect a single key's configuration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get API key details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "pk_abc123def456",
                        "description": "API key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_apikey.APIKey"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Key not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Marks an API key as inactive. The key record remains retrievable via GET for auditing purposes, but all requests using this key will be rejected by the auth middleware with a 401 response.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Deactivate API key",
                "parameters": [
                    {
                        "type": "string",
                        "example": "pk_abc123def456",
                        "description": "API key to deactivate",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: deactivated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/keys/{key}/rotate": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Creates a new API key with the same app name, origins, and rate limit as the old key, then sets a grace period expiration on the old key. Both old and new keys are valid during the grace period, allowing clients to migrate without downtime. After the grace period expires, the old key is automatically rejected.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Rotate API key",
                "parameters": [
                    {
                        "type": "string",
                        "example": "pk_abc123def456",
                        "description": "API key to rotate",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Grace period for old key (Go duration, e.g. 24h, 72h)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.rotateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.rotateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or grace period",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_finish06_drug-gate_internal_apikey.APIKey": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "app_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "origins": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rate_limit": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "No drug found for NDC 99999-9999"
                }
            }
        },
...
        "internal_handler.createKeyRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "origins": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rate_limit": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.rotateKeyRequest": {
            "type": "object",
            "properties": {
                "grace_period": {
                    "type": "string"
                }
            }
        },
        "internal_handler.rotateKeyResponse": {
            "type": "object",
            "properties": {
                "new_key": {
                    "type": "string"
                },
                "old_key": {
                    "type": "string"
                },
                "old_key_expires_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token set via `ADMIN_SECRET` environment variable. Format: `Bearer {secret}`. Required for all `/admin/*` endpoints.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Publishable API key for frontend applications. Create one via `POST /admin/keys`. Include in all `/v1/*` requests.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    },
...
}